refactor: move fhe_types to typing
