+++ fhe_http_core/src/apis/mod.rs
@@ -1,5 +1,4 @@
 pub mod base64;
 pub mod fhe;
 pub mod fhe_ops;
-pub mod fhe_types;
 pub mod http;
