+++ fhe_http_core/src/configs/mod.rs
@@ -1,3 +1,2 @@
-pub mod fhe_types;
-pub mod json;
+pub mod traits;
 pub mod typing;
