+++ fhe_http_core/src/fhe_traits/computable.rs
@@ -1,5 +1,5 @@
-use crate::configs::fhe_types::FheSupportedType;
-use crate::configs::json::FheJsonValue;
+use crate::configs::typing::FheJsonValue;
+use crate::configs::typing::FheSupportedType;
 use crate::fhe_traits::decryptable::Decryptable;
 pub trait Computable {
     fn add(&self, other: &Self, data_type: &FheSupportedType) -> Self;
