+++ fhe_http_core/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod configs {
-    pub mod fhe_types;
-    pub mod json;
+    pub mod traits;
     pub mod typing;
 }
 
@@ -24,7 +23,6 @@ pub mod apis {
     pub mod base64;
     pub mod fhe;
     pub mod fhe_ops;
-    pub mod fhe_types;
     pub mod http;
 }
