+++ fhe_http_core/src/utils/json.rs
@@ -1,4 +1,4 @@
-use crate::configs::json::{FheJsonValue, NorJsonValue};
+use crate::configs::typing::{FheJsonValue, NorJsonValue};
 use crate::fhe_traits::decryptable::Decryptable;
 use crate::fhe_traits::encryptable::Encryptable;
 use crate::utils::base64;
