+++ fhe_http_core/tests/fhe.rs
@@ -1,6 +1,6 @@
 #[cfg(test)]
 mod fhe_tests {
-    use fhe_http_core::configs::json::{FheJsonValue, NorJsonValue};
+    use fhe_http_core::configs::typing::{FheJsonValue, NorJsonValue};
     use fhe_http_core::fhe_traits::decryptable::Decryptable;
     use fhe_http_core::fhe_traits::encryptable::Encryptable;
     use tfhe::{generate_keys, ClientKey, ConfigBuilder, Error};
@@ -52,7 +52,7 @@ mod fhe_tests {
 
     #[test]
     fn serialize_and_deserialize() {
-        use fhe_http_core::configs::fhe_types::FheSupportedType;
+        use fhe_http_core::configs::typing::FheSupportedType;
         use fhe_http_core::fhe_traits::computable::Computable;
         use fhe_http_core::fhe_traits::key_serialize::KeySerialize;
         use tfhe::{set_server_key, ServerKey};
