+++ fhe_http_core/tests/fhe_ops.rs
@@ -1,7 +1,7 @@
 #[cfg(test)]
 mod fhe_tests {
-    use fhe_http_core::configs::fhe_types::FheSupportedType;
-    use fhe_http_core::configs::json::NorJsonValue;
+    use fhe_http_core::configs::typing::FheSupportedType;
+    use fhe_http_core::configs::typing::NorJsonValue;
     use fhe_http_core::fhe_traits::computable::Computable;
     use fhe_http_core::fhe_traits::decryptable::Decryptable;
     use fhe_http_core::fhe_traits::encryptable::Encryptable;
